#1112 remove now-unused file promise code
