+++ allensdk/brain_observatory/ecephys/ecephys_session_api/ecephys_nwb_session_api.py
@@ -6,9 +6,9 @@ import ast
 import pandas as pd
 import numpy as np
 import xarray as xr
+import pynwb
 
 from .ecephys_session_api import EcephysSessionApi
-from allensdk.brain_observatory.ecephys.file_promise import FilePromise
 from allensdk.brain_observatory.nwb.nwb_api import NwbApi
 import allensdk.brain_observatory.ecephys.nwb  # noqa Necessary to import pyNWB namespaces
 from allensdk.brain_observatory.ecephys import get_unit_filter_value
@@ -21,7 +21,7 @@ class EcephysNwbSessionApi(NwbApi, EcephysSessionApi):
 
     def __init__(self,
                  path,
-                 probe_lfp_paths: Optional[Dict[int, FilePromise]] = None,
+                 probe_lfp_paths: Optional[Dict[int, Callable[[], pynwb.NWBFile]]] = None,
                  additional_unit_metrics=None,
                  external_channel_columns=None,
                  **kwargs):
